app/requirements.txt
pandas==1.5.3
numpy==1.26.4
matplotlib==3.7.0
seaborn==0.13.2
streamlit==1.32.2
joblib==1.3.2
scikit-learn==1.4.1.post1
mlflow==2.11.3
category-encoders==2.6.3
configparser==5.2.0
holidays==0.38
psutil==5.9.0
typing-extensions==4.8.0
tensorflow==2.15.0
cffi==1.15.1
cloudpickle==2.2.1
databricks-automl-runtime==0.2.20
defusedxml==0.7.1
graphviz==0.20.3
lightgbm==4.1.0
torch==2.2.2
shap==0.45.0

